decomposed mapping collector
